add comments and remove unnecessary code
