+++ .env.example
@@ -9,7 +9,6 @@ WP_SITEURL="http://${SERVER_NAME}/wp"
 # CDN_URL="http://${SERVER_NAME}:8080"
 
 DB_NAME=justice
-DB_NAME_TEST=justice_test
 DB_HOST=mariadb
 DB_USER=mysql
 DB_PASSWORD=mysql
