+++ config/application.php
@@ -75,7 +75,6 @@ if (env('DB_SSL')) {
     Config::define('MYSQL_CLIENT_FLAGS', MYSQLI_CLIENT_SSL);
 }
 
-
 Config::define('DB_NAME', env('DB_NAME'));
 Config::define('DB_USER', env('DB_USER'));
 Config::define('DB_PASSWORD', env('DB_PASSWORD'));
@@ -83,25 +82,14 @@ Config::define('DB_HOST', env('DB_HOST') ?: 'localhost');
 Config::define('DB_CHARSET', 'utf8mb4');
 Config::define('DB_COLLATE', '');
 
-
 // If the request origin is from a test suite, use the test database.
 $is_test_request = ( isset($_SERVER['HTTP_X_TEST_REQUEST']) && $_SERVER['HTTP_X_TEST_REQUEST'] )
 || ( isset($_SERVER['HTTP_USER_AGENT']) && $_SERVER['HTTP_USER_AGENT'] === 'wp-browser' )
 || getenv('WPBROWSER_HOST_REQUEST');
 
-
+// Set the table prefix based on the request origin.
 $table_prefix =  $is_test_request ? 'test_' : (env('DB_PREFIX') ?: 'wp_');
 
-// echo $table_prefix;
-
-// if (env('DATABASE_URL')) {
-//     $dsn = (object) parse_url(env('DATABASE_URL'));
-
-//     Config::define('DB_NAME', substr($dsn->path, 1));
-//     Config::define('DB_USER', $dsn->user);
-//     Config::define('DB_PASSWORD', $dsn->pass ?? null);
-//     Config::define('DB_HOST', isset($dsn->port) ? "{$dsn->host}:{$dsn->port}" : $dsn->host);
-// }
 
 /**
  * Authentication Unique Keys and Salts
